+++ npm/src/directory-sync/DirectoryConfig.ts
